+++ lib/logger.js
@@ -3,34 +3,7 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
     return (mod && mod.__esModule) ? mod : { "default": mod };
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-const path_1 = __importDefault(require("path"));
-const logger_1 = __importDefault(require("./utils/logger"));
-const fs_1 = require("fs");
-let cwd = process.cwd();
-const requirePaths = [];
-while (cwd !== '/') {
-    requirePaths.push(cwd);
-    cwd = path_1.default.join(cwd, '../');
-}
-const configPaths = requirePaths.map((p) => path_1.default.join(p, 'light.config.js'));
-const config = configPaths.reduce((acc, val) => {
-    let conf = {};
-    if (fs_1.existsSync(val)) {
-        try {
-            console.log('requiring');
-            conf = require(val);
-        }
-        catch (err) {
-            logger_1.default.error(`unable to import light.config.js: ${err}`);
-        }
-    }
-    return Object.assign({}, acc, conf);
-}, {});
+const pino_1 = __importDefault(require("./pino"));
 exports.default = (message) => {
-    if (config.logger) {
-        console.log(`${config.logger} - ${message}`);
-    }
-    else {
-        console.log(message);
-    }
+    pino_1.default.logger.info(message);
 };
