+++ lib/utils/route/logger.js
@@ -3,14 +3,13 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
     return (mod && mod.__esModule) ? mod : { "default": mod };
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-const pino_http_1 = __importDefault(require("pino-http"));
 const bytes_1 = __importDefault(require("bytes"));
+const pino_1 = __importDefault(require("../../pino"));
 const is_prod_1 = __importDefault(require("../is-prod"));
 /* istanbul ignore next */
 if (is_prod_1.default) {
-    const pinoHandler = pino_http_1.default();
     module.exports = (fn) => async (req, res) => {
-        pinoHandler(req, res);
+        pino_1.default(req, res);
         return fn(req, res);
     };
 }
